+++ src/Core/Controller/ControllerCallback.js
@@ -1,6 +1,13 @@
-import Routers from '../Config/Routers';
 import Container from '../Container/Container';
 import Error from '../Logger/Error';
+import Redirect from './Redirect';
+import ResponseData from './ResponseData';
+import View from './View';
+import BaseResponse from '../Controller/Response';
+import { Request, Response } from 'express';
+import config from '../../../config';
+import Template from '../Template/Template';
+import ConfigView from '../../App/Config/View';
 
 /**
  * @class ControllerCallback
@@ -12,10 +19,9 @@ class ControllerCallback {
      * @param {string} controller
      * @param {string} fn
      * @param {{}} additionalData
-     * @param {Routers.response} routersResponse
      * @return {Function}
      */
-    static call(controller, fn, additionalData, routersResponse) {
+    static call(controller, fn, additionalData) {
         return async (req, res, next) => {
             try {
                 let controllerInstance = null;
@@ -42,7 +48,7 @@ class ControllerCallback {
                 } else {
                     returnedData = data;
                 }
-                await routersResponse(req, res, returnedData);
+                await ControllerCallback.response(req, res, returnedData);
             } catch (e) {
                 Error.create('error', e.stack);
 
@@ -55,6 +61,43 @@ class ControllerCallback {
                 }
             }
         };
+    }
+
+    /**
+      *
+      * @param {Request} req
+      * @param {Response} res
+      * @param {ResponseData|View|Redirect} returnedData
+      */
+     static async response(req, res, returnedData) {
+        if (returnedData == undefined) {
+            res.status(400).send('Unexpected Error, Method didnt return anything');
+        }
+
+        let response = null;
+        if (returnedData instanceof BaseResponse) {
+            response = await returnedData.send();
+        }
+
+        if (response instanceof ResponseData) {
+            res.status(response.code).json(response.data);
+        }
+
+        if (returnedData instanceof View) {
+            if (returnedData.type == 'html') {
+                res.send(returnedData.view);
+            }
+            if (returnedData.type == 'view') {
+                res.render(returnedData.view, {...returnedData.data, ...Template(), ...ConfigView.hook()});
+            }
+            if (returnedData.type == 'sendFile') {
+                res.sendFile(config.sourcePath + '/App/Views/' + returnedData.view);
+            }
+        }
+
+        if (returnedData instanceof Redirect) {
+            res.redirect(returnedData.route);
+        }
     }
 }
