+++ src/Core/Libraries/DateFormat.js
@@ -39,6 +39,14 @@ class DateFormat {
     static getCurrentDate(format) {
         return moment().format(format);
     }
+
+    /**
+     * 
+     * @param {Date} date 
+     */
+    static databaseDate(date){
+        return moment(date).format('YYYY-MM-DD hh:mm:ss');
+    }
 }
 
 export default DateFormat;
